backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
httpx==0.27.2
SQLAlchemy==2.0.36
asyncpg==0.29.0
python-dotenv==1.0.1
tenacity==9.0.0
